git-svn-id: http://www.bci2000.org/svn/trunk@4446 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/core/Operator/OperatorLib/ObjectTypes/Watches.h
@@ -65,7 +65,7 @@ class Watch
     Set( const Set& );
     ~Set();
     
-    int Size() const { return size(); }
+    int Size() const { return static_cast<int>( size() ); }
     bool Empty() const { return empty(); }
     
     Watch* First() const;
